Separate logic into files
