+++ server/src/server/handler.rs
@@ -1,106 +1,18 @@
 use actix::prelude::*;
-use rand::{self, rngs::ThreadRng, Rng};
-use serde::{Deserialize, Serialize};
+use rand::rngs::ThreadRng;
+use rand::Rng;
 use std::collections::HashMap;
 use tracing::info;
 
-#[derive(Message)]
-#[rtype(result = "()")]
-pub struct Message(pub String);
-
-#[derive(Message)]
-#[rtype(usize)]
-pub struct Connect {
-    pub addr: Recipient<Message>,
-}
-
-#[derive(Message)]
-#[rtype(result = "()")]
-pub struct Disconnect {
-    pub id: usize,
-}
-
-#[derive(Message, Deserialize, Serialize)]
-#[rtype(result = "()")]
-pub struct ClientMessage {
-    pub from_user: usize,
-    pub to_user: usize,
-    pub msg: String,
-}
-
-impl ClientMessage {
-    pub fn new(data: &str) -> Self {
-        serde_json::from_str(data).unwrap()
-    }
-}
-
-#[derive(Message)]
-#[rtype(result = "()")]
-pub struct CommunicateUser {
-    pub ws_id: usize,
-    pub user_data: String,
-    pub comm_type: CommunicationType,
-}
-
-pub enum CommunicationType {
-    SendUserData,
-    CreateNewUser,
-    UpdateUserIDs,
-    UpdateName,
-    UpdateImageLink,
-    ReconnectUser,
-}
-
-#[derive(Message)]
-#[rtype(result = "()")]
-pub struct Join {
-    pub id: usize,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct UserData {
-    id: usize,
-    name: String,
-    image_link: Option<String>,
-}
-
-impl UserData {
-    fn new(data: String) -> Self {
-        serde_json::from_str(&data).unwrap()
-    }
-
-    fn to_json(&self) -> String {
-        serde_json::to_string(self).unwrap()
-    }
-
-    fn update_id(self, id: usize) -> Self {
-        UserData {
-            id,
-            name: self.name,
-            image_link: self.image_link,
-        }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub struct WsData {
-    user_id: usize,
-    ws_id: usize,
-}
-
-impl WsData {
-    fn new(user_id: usize, ws_id: usize) -> Self {
-        WsData { user_id, ws_id }
-    }
-}
+use crate::server::{Message, UserData, WsData};
 
 #[derive(Debug)]
 pub struct ChatServer {
     // {session id: {user id this session belongs to, recipient}}
-    sessions: HashMap<usize, (usize, Recipient<Message>)>,
+    pub sessions: HashMap<usize, (usize, Recipient<Message>)>,
     user_data: HashMap<usize, UserData>,
     user_session: HashMap<usize, Vec<WsData>>,
-    rng: ThreadRng,
+    pub rng: ThreadRng,
 }
 
 impl ChatServer {
@@ -113,10 +25,8 @@ impl ChatServer {
             rng: rand::thread_rng(),
         }
     }
-}
 
-impl ChatServer {
-    fn send_message(&self, message: &str, from_user: usize, to_user: usize) {
+    pub fn send_message(&self, message: &str, from_user: usize, to_user: usize) {
         info!("Sending message from {} to {}", from_user, to_user);
         if let Some(receiver_ws_data) = self.user_session.get(&to_user) {
             for i in receiver_ws_data {
@@ -135,14 +45,14 @@ impl ChatServer {
     }
 
     /// Sends the WS Connection ID to the client
-    fn send_session_id(&self, id: usize) {
+    pub fn send_session_id(&self, id: usize) {
         if let Some((_, receiver_ws)) = self.sessions.get(&id) {
             receiver_ws.do_send(Message(format!("/update-session-id {}", id)))
         };
     }
 
     /// Creates a new user and allocates necessary data to communicate with it
-    fn create_new_user(&mut self, ws_id: usize, other_data: String) {
+    pub fn create_new_user(&mut self, ws_id: usize, other_data: String) {
         let user_id = self.rng.gen::<usize>();
         info!(
             "Creating new user on ws_id {} with user id {user_id}",
@@ -166,7 +76,7 @@ impl ChatServer {
     }
 
     /// Allocates necessary data to communicate with a previously deleted user
-    fn reconnect_user(&mut self, ws_id: usize, user_id: usize, other_data: String) {
+    pub fn reconnect_user(&mut self, ws_id: usize, user_id: usize, other_data: String) {
         info!("Reconnecting with user with id {} on ws {ws_id}", user_id);
         let user_data = UserData::new(other_data).update_id(user_id);
         self.user_data.insert(user_id, user_data);
@@ -185,7 +95,7 @@ impl ChatServer {
     }
 
     /// Sends user profile data to the client
-    fn send_user_data(&mut self, ws_id: usize, id: usize) {
+    pub fn send_user_data(&mut self, ws_id: usize, id: usize) {
         info!("Sending user data of with id {}", id);
         if let Some(data) = self.user_data.get(&id) {
             let user_data = data.to_json();
@@ -196,7 +106,7 @@ impl ChatServer {
     }
 
     /// Used to keep track of active user ws connections
-    fn update_ids(&mut self, ws_id: usize, user_id: usize, client_id: usize) {
+    pub fn update_ids(&mut self, ws_id: usize, user_id: usize, client_id: usize) {
         if let Some(entry) = self.sessions.get_mut(&ws_id) {
             let (session_user_id, _) = entry;
             *session_user_id = user_id;
@@ -212,7 +122,7 @@ impl ChatServer {
     }
 
     /// Updates user name
-    fn update_user_name(&mut self, ws_id: usize, new_name: String) {
+    pub fn update_user_name(&mut self, ws_id: usize, new_name: String) {
         let user_id = &self.sessions[&ws_id].0;
         info!("Updating name of user {} to {new_name}", user_id);
 
@@ -234,7 +144,7 @@ impl ChatServer {
     }
 
     /// Updates image link of a user
-    fn update_user_image_link(&mut self, ws_id: usize, new_link: String) {
+    pub fn update_user_image_link(&mut self, ws_id: usize, new_link: String) {
         let user_id = &self.sessions[&ws_id].0;
         info!("Updating image link of user {} to {new_link}", user_id);
 
@@ -251,69 +161,6 @@ impl ChatServer {
                         }
                     }
                 }
-            }
-        }
-    }
-}
-
-impl Actor for ChatServer {
-    type Context = Context<Self>;
-}
-
-impl Handler<Connect> for ChatServer {
-    type Result = usize;
-
-    fn handle(&mut self, msg: Connect, _: &mut Context<Self>) -> Self::Result {
-        let id = self.rng.gen::<usize>();
-        self.sessions.insert(id, (0, msg.addr));
-        self.send_session_id(id);
-        id
-    }
-}
-
-impl Handler<Disconnect> for ChatServer {
-    type Result = ();
-
-    fn handle(&mut self, msg: Disconnect, _: &mut Context<Self>) {
-        info!("WS Session {} disconnected", msg.id);
-        // TODO remove user details from user_data and sessions
-        // Needs to save each sessions owner/client id perhaps?
-        self.sessions.remove(&msg.id);
-    }
-}
-
-impl Handler<ClientMessage> for ChatServer {
-    type Result = ();
-
-    fn handle(&mut self, msg: ClientMessage, _: &mut Context<Self>) {
-        self.send_message(&msg.msg, msg.from_user, msg.to_user);
-    }
-}
-
-impl Handler<CommunicateUser> for ChatServer {
-    type Result = ();
-
-    fn handle(&mut self, msg: CommunicateUser, _: &mut Context<Self>) {
-        match msg.comm_type {
-            CommunicationType::SendUserData => {
-                self.send_user_data(msg.ws_id, msg.user_data.parse().unwrap())
-            }
-            CommunicationType::CreateNewUser => self.create_new_user(msg.ws_id, msg.user_data),
-            CommunicationType::UpdateUserIDs => {
-                let data: Vec<&str> = msg.user_data.split(' ').collect();
-                self.update_ids(
-                    msg.ws_id,
-                    data[0].parse().unwrap(),
-                    data[1].parse().unwrap(),
-                );
-            }
-            CommunicationType::UpdateName => self.update_user_name(msg.ws_id, msg.user_data),
-            CommunicationType::UpdateImageLink => {
-                self.update_user_image_link(msg.ws_id, msg.user_data)
-            }
-            CommunicationType::ReconnectUser => {
-                let data: Vec<&str> = msg.user_data.splitn(2, ' ').collect();
-                self.reconnect_user(msg.ws_id, data[0].parse().unwrap(), data[1].to_string());
             }
         }
     }
