+++ server/src/server/mod.rs
@@ -1,0 +1,8 @@
+mod handler;
+mod server;
+
+pub use handler::ChatServer;
+pub use server::{
+    ClientMessage, CommunicateUser, CommunicationType, Connect, Disconnect, Message, UserData,
+    WsData,
+};
